Create Vacuum_Cleaner_Agent_.ipynb
This is code for Vacuum Cleaner Agent to clean the room 
The room is in the matrix form of [10*10]. Initially some room is clean and some is dirty denoting clean room by 0 and dirty room by 1. Scanning is start from first row of first room when reached to last room of first row, it move to last room of second row to reduce time to come in first room of second row which is very effective in real life. While scanning if it found clean room if move forward and if dirty room found it clean and then move forward at last when all room become clean it stop.